requirements.txt
pillow
pyzbar
protobuf